Moves privacy policy to own folder.
