+++ src/frontend/screens/PrivacyPolicy/PointContainer.tsx
